requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
geographiclib==2.0
geopy==2.4.0
h11==0.14.0
httptools==0.6.1
idna==3.4
mysql-connector-python==8.2.0
numpy==1.26.2
protobuf==4.21.12
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
PyYAML==6.0.1
scipy==1.11.4
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
